Ocean: revamp setoracledata indexer (#2987)
* infer

* revamp setoracle index

* forward aggregate value

* backward aggregate value

* invalidate

* fmt

* note the changes

* rm unuse -.-

* revert -

* div correction

* revert price_ticket by_key

* revert not put at oracle_price_agg by_key

* fmt

+++ lib/ain-ocean/src/indexer/poolswap.rs
@@ -74,7 +74,7 @@ fn index_swap_aggregated(
                 .unwrap_or(dec!(0));
 
             let aggregated_amount = amount
-                .checked_add(Decimal::from(from_amount / COIN))
+                .checked_add(Decimal::from(from_amount) / Decimal::from(COIN))
                 .ok_or(Error::OverflowError)?;
 
             aggregated.aggregated.amounts.insert(
@@ -137,7 +137,7 @@ fn invalidate_swap_aggregated(
                 .unwrap_or(dec!(0));
 
             let aggregated_amount = amount
-                .checked_sub(Decimal::from(from_amount / COIN))
+                .checked_sub(Decimal::from(from_amount) / Decimal::from(COIN))
                 .ok_or(Error::UnderflowError)?;
 
             aggregated.aggregated.amounts.insert(
